fix: hex independency infra for model3d

+++ app/Infrastructure/projectRepository.ts
@@ -30,22 +30,29 @@ class ProjectRepository implements IProjectRepository {
   }
 
   async getById(id: string): Promise<Project | null> {
-    return await this.db.Projects.findById(id);
+    const result = await this.db.Projects.findById(id);
+    return result ? this.toDomainEntity(result) : null;
   }
 
   async create(project: Project): Promise<Project> {
     const newProject = new this.db.Projects(project);
-    return await newProject.save();
+    const result = await newProject.save();
+    return this.toDomainEntity(result);
   }
 
   async update(id: string, project: Project): Promise<Project | null> {
-    return await this.db.Projects.findByIdAndUpdate(id, project, {
+    const result = await this.db.Projects.findByIdAndUpdate(id, project, {
       new: true,
     });
+    return result ? this.toDomainEntity(result) : null;
   }
 
   async deleteById(id: string): Promise<void> {
-    await this.db.Projects.deleteOne({ _id: id });
+    const projectToDelete = await this.db.Projects.findById(id);
+
+    if (projectToDelete) {
+      this.db.Projects.deleteOne({ _id: id });
+    }
   }
 }
 export { ProjectRepository };
